Add emergencies table in country operation tab
